Separate grpc-protocol project.
