+++ grpc-protocol/build.gradle
@@ -1,0 +1,1 @@
+// Tests for this module are in grpc/src/test/java/com/linecorp/armeria/common/grpc/protocol
