+++ settings.gradle
@@ -9,6 +9,7 @@ includeWithFlags ':bom',                       'bom'
 includeWithFlags ':core',                               'java', 'publish', 'shade', 'trim'
 includeWithFlags ':rxjava',                             'java', 'publish', 'relocate'
 includeWithFlags ':grpc',                               'java', 'publish', 'relocate'
+includeWithFlags ':grpc-protocol',                      'java', 'publish', 'relocate'
 includeWithFlags ':jetty',                              'java', 'publish', 'relocate'
 includeWithFlags ':kafka',                              'java', 'publish', 'relocate'
 includeWithFlags ':logback',                            'java', 'publish', 'relocate'
